+++ zpack/zpack.h
@@ -52,7 +52,7 @@ namespace zp
 
 typedef unsigned char u8;
 typedef unsigned short u16;
-typedef unsigned long u32;
+typedef unsigned int u32;
 typedef unsigned long long u64;
 
 const u32 MAX_FILENAME_LEN = 1024;
